chore: lawbottomsheet props 계층 분리
