+++ src/features/laws/ui/LawBottomSheet.tsx
@@ -2,16 +2,7 @@ import { Sheet as ModalSheet } from 'react-modal-sheet';
 import { Sheet } from '@/shared/ui/sheet/Sheet';
 import { CheckList } from '@/shared/ui/check-list/CheckList';
 import { LAW_LIST } from '@/features/laws/constants/law-list';
-
-export type LawBottomSheetProps = {
-  onClose: () => void;
-  isOpen: boolean;
-  onClickPrimaryBtn: () => void;
-  agreements: { [key: string]: boolean };
-  onCheck: (id: string, checked: boolean) => void;
-  onClickLawDetail: (id: string) => void;
-  allAgreed: boolean;
-};
+import { LawBottomSheetProps } from '@/shared/ui/sheet/bottomSheet';
 
 export function LawBottomSheet({
   isOpen,
